# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "告警管理后端服务 API 文档，支持告警接收、静默、抑制、通知、多租户管理、RBAC 权限控制等功能。",
        "title": "api-server — 告警管理 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "paths": {
...
        "/api/v1/alertChannel": {
            "get": {
                "description": "获取所有 AlerChannel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerChannel 管理"
                ],
                "summary": "获取所有 AlerChannel",
                "parameters": [
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page 从 1 开始, 表示第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "PageSize 最大值 100，默认值 20, 表示每页多少条数据",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "feishuApp",
                            "feishuBoot",
                            "webhook",
                            "email"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建 AlerChannel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerChannel 管理"
                ],
                "summary": "创建 AlerChannel",
                "parameters": [
                    {
                        "description": "创建请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertChannelCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertChannel/:id": {
            "get": {
                "description": "查询 AlerChannel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerChannel 管理"
                ],
                "summary": "查询 AlerChannel",
                "parameters": [
                    {
                        "description": "查询请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新 AlerChannel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerChannel 管理"
                ],
                "summary": "更新 AlerChannel",
                "parameters": [
                    {
                        "description": "更新请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertChannelUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 AlerChannel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlerChannel 管理"
                ],
                "summary": "删除 AlerChannel",
                "parameters": [
                    {
                        "description": "删除请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertHistory": {
            "get": {
                "description": "获取所有 AlertHistory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertHistory 管理"
                ],
                "summary": "获取所有 AlertHistory",
                "parameters": [
                    {
                        "type": "string",
                        "name": "alertName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "alertSendRecordId",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "endsAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fingerprint",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "instance",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "labels",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page 从 1 开始, 表示第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "PageSize 最大值 100，默认值 20, 表示每页多少条数据",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "severity",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "alertname",
                            "fingerprint",
                            "starts_at",
                            "ends_at",
                            "severity",
                            "instance"
                        ],
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "startsAt",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "resolved",
                            "firing"
                        ],
                        "type": "string",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/alertHistory/:id": {
            "get": {
                "description": "使用 ID 查询告警历史详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertHistory 管理"
                ],
                "summary": "查询 AlertHistory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "告警历史ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新 AlertHistory 状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertHistory 管理"
                ],
                "summary": "更新 AlertHistory 状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "告警历史ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AlertHistoryUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.AlertChannelCreateRequest": {
            "type": "object",
            "required": [
                "config",
                "name",
                "status",
                "type"
            ],
            "properties": {
                "aggregationStatus": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 15
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "feishuApp",
                        "feishuBoot",
                        "webhook",
                        "email"
                    ]
                }
            }
        },
        "types.AlertChannelUpdateRequest": {
            "type": "object",
            "required": [
                "aggregationStatus",
                "config",
                "id",
                "status",
                "type"
            ],
            "properties": {
                "aggregationStatus": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "feishuApp",
                        "feishuBoot",
                        "webhook",
                        "email"
                    ]
                }
            }
        },
        "types.AlertHistoryUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "types.IDRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}